Add more typical main block
This is simply a convention which allows you to be freer to import the
FaceBase_Launch.py file into another script without it launching the UI.

Only when the FaceBase_Launch.py script is executed directly will the
built-in __name__ variable be defined as "__main__". During import it
will be the name of the file itself and so won't trigger this
if-statement.