Add simple status bar message logger
We create a message logger class which wraps the status bar from the
main application window and exposes a simple message logging interface
to the rest of the code.

We pass it to the RigGraphicsView and use it to report one of the errors
in a simple string manner. Should be improved to use a coloured error
panel. Most of the change in RigUIControls is a space change due to
reindenting the function after moving the error check and exit to the
top.