Centralise some of the stylesheet reading
We switch to reading the stylesheet information up front and complaining
if it isn't there. We then pass this into the classes which care about
it rather than reading it in the instance set up code.

We also move to having a main() function just to make it easier to exit
out of the set up code if needed. We also return an appropriate value on
error or successful exit. This isn't really important for a GUI
application but is good practice.