Saving and Reading of SuperNodes and Skinning Complete
This is a milestone from the point of view in practice all saving and
loading of the functonality is now possible. This means that complex faces
can now be stored and reloaded.

Restrictions are that nodes can currently only be skinned properly to one
superNode. Yuo can skin to more, but the values will fight overriding
eachother in a way that is very confusing. Consider adding multiple
superNode influences to SkinningInfo. Discussion on how to do this.

In this commit there are also a number of doc strings added to try and
make it clearer the purpose of each class. doc strings for classes will be
the priority, followed by doc strings for complex or unintuitive methods.