Additional Docstrings added for classes in other files
DocStrings added to describe the main purpose and interaction of all
classes in the FaceBase_Launch.py, FileControl.py, RigStore.py
Utilities.py adn Widgets.py. Some more details still need to be added to
the main subclassed RigGraphicsView.

Once all of this is done then a second wave of Docstrings will be
implemented for the major functions within each class. Currently only a
handful of functions are commented in any detail.