Switch to ItemFactory from eval() method
Eval should be considered dangerous so we switch to a cleaner table
lookup mechanism for figuring out what classes to create from the text
mime data from the drop event.

We add .name class level values to the widget classes and construct a
table of name -> constructor pairs. This keeps the look up name nice and
close to the class definition.

We also change to using the .name attributes instead of the strings when
creating the DragItem controls to make sure everything stays in sync.