Neatened up WireGroup Curve draw - additional method
This still is in huge need of optimising, but an extra method has been
added to the node class that allows us to call the update from one method
call. This has allowed a couple of lines of code to be simplified down to
one call in a number of areas.