Improved curve performance
We change from rebuilding the curve every time anything on it changes to
instead marking it dirty and only rebuilding it in the paint method if
it is marked as dirty.

The dirtying operation is very quick and the build operation only has to
be done once per draw rather than every time anything changes along the
curve. This makes it significantly more performant.

We rename curveUpdate to dirtyCurve to reflect the change of purpose.