Functionality for Node Merging into another Wiregroup working
Now a node can be merged across into another Wiregroup by clicking on the
node and pressing CTRL and dragging the node over the node you want to
merge to. The node you want to merge to will glow white when you are over
it. Releasing the mouse will merge it.

Note that it is the targetNode that you are moving to that stays in the
scene, the node you are moving is destroyed in the merge. The targetNode
then takes its place in the node/pin/pinTie arrays of both WireGroups.

Once the node takes its place in the arrays of the WireGroup the current
Wiregroup RigCurve is destroyed, and the references to it are stripped
from the Nodes. A Completely new Curve is the built from scratch for the
wiregroup using the new arrays.