Storage is now working for reference Nodes/Pins for Node Merges
Storage now works for Nodes and Control Pins. An extra attribute is
included in the storage, which indicates if the Node/Pin is native or
referenced from another WireGroup. If native, the node/pin is constructed
on "read"  as normal. If referenced then a NodePinReference() object is
created which stores the wiregroup name and index required to find the
node/pin and drop it in later.

In turn this means that the WireGroup curves cannot be generated until
until the references are resolved, with the referenced node being dropped
in the slef.nodes and self.pins lists. Once this is implemented then the
curves can be drawn.