Store and Load of Node Merging is now supported
Now nodes/pins can be merged into other wiregroups. The results can be
stored and reloaded in through the XML face files. Currently multiple
levels of merging are allowed, so a node can merge into more than one
WireGroup. On the most part this is stable, but there does seem to a
couple of very complex cases where the read, after storing fails.
Investigate this. I think it will be bullet proof is you do try and merge
a node that is already merged.

The storage is done by checking whether a node/pin is native to the
current WireGroup being stored. If it is native (as in created within the
Wiregroup initially) then it is stored with an attribute (state) as
native. If it is not native (has been merged in from another Wiregroup)
then is stored as state = reference.

This is picked up by the read method on the Wiregroup at loading. If it
finds a native node/pin then it rebuilds it normally. If it finds a
reference, then is creates a instance of a NodePinReference and stores it
in place of a node or pin in the WireGroup lists.

Once all Wiregroups are built, then these references are processed and
correct nodes/pins from neighbouring Wiregroups are substituted in. Once
this process is resolved, then the final Wiregroup curves are built.

There are lots of doc strings in the methods to help describe this process