Basic Control Scaling Implemented along the Top Tool Bar
In Widgets a class called ControlScale has been created. Into this can be
loaded a scene, a QSlider, and then a list of objects. The ControlScale is
passed to the rigGraphicsView, so when objects are selected then they can
be filtered out and scaling can occur on the appropriate controls in the
scene.

At the moment this is mostly implemented, with scaling on individual
objects working. However, scaling on muliple selections and scaling on
all controls of the same type has not been implemented.

Scaling needs to be passed down to the pins, so ControlPins are scaled
appropriately as well.

There is a checkbox on the toolbar called "Scale Matching Items". This has
not been implemented, but when it is then it will allow the user to scale
all controls in the scene to the same scale. So for example you could
shrink all of the guidemarkers...