Implementation of all scaling to single and multiple controls
Nearly all issues tidied up with regards scaling.

The ControlScale() checks that we have only one type of control selected.

The RigGView MouseReleaseEvent handles the mouse button release, and then
processes the controlList in case scale using the slider is about to take
place.

Scale now works on a selection or single Item. If the checkbox "Scale
Matching Items" is checked then the scale will apply to all scene Items of
the same type as the first item selected. This is used for more global
changes.

Fixed one tiny bug with the store method of supernodes, where an
additional arguement had to be passed to the pin.store()

Scaling still needs to be passed on to ControlPins

There is now bug with using Constraint Items, which conflicts with the new
scaling methods. This needs resolving.