Scaling cleaned up by filtering Items correctly
Checks are now in place for to check the first item in the control list in
mouseReleaseEvent (RigGraphicsView). If the first Item is of the wrong
type then the controllist is marked as not valid.

A bug has also been fixed with the loading (read) of superNode groups
losing the links to their constraintItems. This now loads properly.