Deletion implemented for WireGroups
On pressing the 'delete' key, the RigGraphicsView picks up the in the
keypressed method and checks the type of the item. This is done with if
statements, but could be neatened to polymorphism around a clear() method.
eaten when time allows. The Wiregroup is then removed from the
self.wiregroups list on the RigGraphicsView