Naming on Drop for SuperNodeGroups
Adjusted the processDrop() method in the RigGraphicsView to detect the
drop of a SuperNodeGroup. The self.dragItem is the pin. So the pins Group
is tested (getGroup()) to see if it is a member of a superNodeGroup (it
might be a memeber of a WireGroup). If it is a memeber of a superNodeGroup
then we can use a QInputDialog to grab and check some text to apply as the
name. The name has to be unique. If the naming is cancelled, then the
SuperNodeGroup is deleted...