Tidying User Experience - sorting Right Click Orders etc
For RC context menu in the main RigGraphicsView, the nodes are now sorted
into order according to a list (sortMenuItem()). This allows the most
useful RC menus to appear even if they counter the item Z height in the
scene compared to other items.

Additionally the bounding boxes of nodes and ControlPins were corrected
to be neater and more representative of the actual shape.

ControlPin drawing was reduced to an unfussy simple cross to declutter the
scene.

ItemIsMoveable, ItemIsSelectable and a few other flags were turned off on
ConstraintLines to stop them fighting for selection with other Nodes that
were drawn before the Node and Pin with  the ConstraintLine. Being drawn
before meant that they were lower in Z, and therefore other Nodes were not
selectable, because of the constraintLine, even through you would never
want to select the constraint line.