Action menu added to the top of UI
Action menu includes the option to reflect Guide Markers. This might get
upgraded to reflect any type of control or item in the RigGraphicsView.

The "Clear All" option was also added top the menu with a separator. This
has allows the user to clear the RigGraphicsView from the action menu. In
the clear() method a query input has been added, so the user is presented
with a messagebox, when they try and clear in this way.