Filter Option Tool bar renamed to Quick Tools
Rename has occured since the tool bar has filled up with much more
generally useful tools. Space has also been added to the start of the
toolbar using a Qlabel widget with just a load of spaces as the name. This
is just make start point of the tool bar line up to a more athestically
pleasing place.