Basic Load in Maya 2014 PySide Achieved
The basic load works with the UI loading and being correctly parented to
the Maya Window. However, all the relative referencing to the UI images
has failed, since the execution seems to think that the current directory
is complete different...

Therefore implementing fuller paths to the image locations using the
directory for __file__. This might not be the neatest way of doing it!

This commit is with the basic UI loading, but the images failing. Can
roll back to here if a more cunning plan is formulated.