Creation Menus Icons now Loading (CSS path fix) but not working
The fix had to occur in the validImageFile method of the DragItemButton
series in Widgets. Before the paths were relatives referenced from the
current directory, which was where the code was. Maya has it own idea of
the current directory, so the __file__ variable was brought into play.

This kicks up loads of \ on windows, which you are apparently able to
convert to / for the CSS paths using elegant os.path and posixpath
methods. Bollocks to that, it doesnt work, so it currently works with nice
direct strong replacement for \ and /