Adding basic Structire for Data Processing
New file added called dataProcessing.py

The structure is that there is a DataProcessor() class that is
responsible for pooling all the interesting data changes that occur.

The DataProcessor collects all the little DataBundles(). These bundles
are little bundles of information that are attached to each Node and
SuperNode. They register the positional changes and translate that into
outputting a useful float data value that we can then attach out to
anything we want.

At the moment this is just a basic implementation to pass the
DataProcessor down to the node level. Now build by testing and activating
functionality