DataBundles now active and reporting data on basic position Changes
The DataBundles and now created and attached to the node when the node is
created. This is done internally in the node when it is created and the
dataProcessor is passed to the constructor
(self.dataProcessor.attachBundle(node))

The additional code has been inserted in the itemChanged() method for the
node and superNode to update the dataBundle given the new position. The
result is the setX() and setY() methods that are called and spit out a
useful float value for each axis.

Currently constraint Items are not considered at all, that needs to be
implemeted next to manage the float output to be scaled appropriately to
the regions of the constraint Items.

Have to check that the superNode Item which inherits the Node Item is not
processing data twice in the itemChanged() method