Good progress on Constraint Ellipse callibration for DataBundles
calibrateDataBundle() method has been added to the ConstraintEllipse, and
this will be added to all other constraint Items. This sets the maxX,
minX, maxY, minY of the dataBundle to align with the shape of the
constraintItem. When the constraint Item is created this method is called
in the ControlPin setConstraintItem method.

It is also called in the redraw() method of the constraintItem itself so
dataBundle calibration is updated whenever the constraintItem shape is
adjusted