Tweak to movement of Nodes and GuideMarkers
Movement of these GraphicsItems are now updated, so that on
MousePressEvent, then position of the item is snapped centrally to the
mouse cursor positions, so you are dragging the item around directly from
its centre.

Nodes had their MousePressEvent extended. PinTie is also updated in this
method.

GuideMarkers had to have the MousePressEvent methods implemented solely
for this purpose