Addition of Scene Controller builder
A Scene controller can now be added. This is done by building a locator
into the Maya scene. Code is separated into mayaData.py so this module can
be easily replaced by other modules for other 3D packages.

Methods have been added to MayaData to build a scene Controller, check
whether a name exists in the Maya scene (isNameUnique). Further methods
have been added to add an Attribute to a node, and to check if an
attribute of a certain name already exists on a Node.

In ControlItems.py adjustments have been made to the WireGroup and Node
Classes. Node Method to identify the attribute name that would relate it
to the SceneController has been added, using the WireGroup Name and the
Node index. The DataBundle now has access to the SceneController, the
correct float values for the node position, and the attribute name it
should correspond to. This means that everything should be able to be
linked up to control the extra attribute value on the Scene Controller
when it is constructed.