More details added to Maya CreateSceneControl
For the animatronic system, this seems to the the fastest way to do
things. We need a scene controller that can have clearly labeeled extra
attributes that are the driving values behind the movement. In this commit
the mayaData createSceneController method has been extended to tidy up and
hide unwanted attributes and then displace a "HappyFace" attribute which
is a True Boolean that is locked to reflect the start of the code rig
Attributs.

Methods "addAttr" and "addTitleAttr" were added to allow this to be done
easily and to add more key attributes later.