DataBundles - Host Name added - Methods UI controls
dataProcessor methods to collect and check all the wiregroups and
supernodegroups in the rigGV are now implemented. The dataprocessor has
the rigGV passed to it, so that it can reference everything it needs.

methods added to cycle through wiregroups and superNodegroups to check the
databundles are correct. check not fully implemented.

Decided that each databundle shoudl have a hostname, which is the name of
the wiregroup or superwiregroup that its associated node comes from. This
is initially blocked in, but currently not working... look into why.