Tweat to setHost Names correctly in WireGroups and SuperNodeGroups
The nodes and supernodes now have the entire wiregroup or SuperNodeGroup
set on a Group rename, rather than just the name being set as string. But
setting the entire group the data is updated, and the setHostName methods
are called to update the dataBundles. Now working for both group types.
Get that there isnt conflict in other areas.

Consider changing Nodes to have a "Group", so you can set the group
without caring whether it is a node group or a WireGroup...