Basic SceneLinkTabW added to bottom Data Tables section
The lower UI fock widget was just for skinning information. This has been
expanded so that it incorporates other data tables by using a tabWidget
for multiple tabs.

The SceneLinkTabW has be added to Widgets.py. Currently setup with a
simple link of adata added.

Need to create a new class to pipe all the channel data into. Store these
channels in the main dataProcessor, and then use them to move changes
easily backwards and forwards from any table changes.