Basic structure for SceneLinkTabW implemented
The QTableWidget now basically populates by collecting the nodes,
associated databundles and then from them a list of all the active
AttributeControllers.

The data is feeding into the table well, but some adjustments need to be
made. The NodeID name is incomplete, and the row number is not updating
properly, unless the update button is clicked twice. Tweak these, and then
the RC menus need to be added so that the rest of the data can be added
smoothly.