Filters placed on servo angle min and max table inputs
A bit of a hacky solution implemented to call a function on "itemchanged"
signal, which checks the column, checks whether the data is a number, and
if it is allows it, if not, then sets the angle to 0 or 180. would be nice
to just ignore the value change, and return it to whatever it was. Fix
should be simple when the trick is figured.

When populating both data tables signals are now blocked, to limit
expensive computational rubbish firing off all over the place.

Next implement node highlighting in the graphics view, when you select a
row of the data table.

Then look into linking to scene Nodes...