Node and SuperNode Highlighting is now implemented from SceneLinkTabW
When you click the row the corresponding node/superNode now highlights.
This was done by adding the node itself to the attributeConnector for easy
reference. The isHighlighted attribute on the Node and SuperNode was then
assigned True/False depending on the row selection.

The mouseMoveEvent was tidied in the graphicsView,since it was excessively
calling "setHighlight" for each Node.

Merging Nodes from WireGroups was updated, to account for getGroup
changes. Merged Nodes still need to be filtered for creating a final list
of activeAttributeControllers, since duplicates will currently occur