Scene Link Node Finalised and Scene Link Attribute Implemented
New sceneNodeAttrContextMenu has been implemented. It first checks the
validity of the referenced scene node, then if the node exists, it
provides a list of possible attributes that can be connected to.

Can choose to detach attribute and Node.

Need to implement Node and Attribute checker. Where is both are the same,
then another AttributeConnector cannot have the same combination of Node
and attribute, so the attribute will be reset to None if a clash occurs