sceneNodeActive and basic Maya Scene Node wiring Implemented
Now when the node and the attribute are valid, a sceneNodeActive attribute
is set to True in the Attribute Connector. This then allows in the
DataBundle setX and setY methods for data to be passed through to the
scene Node.

This is working, but there is lag, possibly due to the print statements.
There is also problems with the positioning and clicking not quite being
right. Look into this for the Node movement. Something seems to be going
wrong with the mouseclickEvent too. Maybe it is not correctly returning,
and ruining the click state. Investigate.