FocusInEvent and FocusOut Event implemented for main Rig GV
This is a hack that has been put together. If alt is pressed in the GV and
then focus is lost, then the GV is left in a frozen state where certain
items are unselectable and unMoveable. This is because the release of alt
(keyReleaseEvent) is not detected, since the focus of the window is lost.

Hack fix, is to check whether items are currently being frozen for panning
in the focusOutEvent in the GV. If they are then we need to release them
before the window loses focus. Uses panRelseaseSelectedItems()