Min and Max Scale Limits added to NodeLink on SceneLinkTabW
Min and Max setup tweaked so that the correct behaviour occurs in x and y
directions.

Improved checkDigit method written, to properly recognise all floats
including decimals.

SceneLinkTabW updateSceneLinkOutPutData method updated to properly update
the attributeConnector, this is followed buy a self.populate() call to
bring TableWidget up to date.

RandomNode: The connection to the scene Node seems fine. There is a little
lag when the superNode is constrained to an area. There is also a lot more
lag when the node that is being controlled is also selected in the
viewport.... might have to manipulate selection when running the code...