Stopped keypresses in RigGraphicsView propagating up to Parent
Shortcut keypresses were propergating up to Maya since in the
RigGraphicsView the keypressevent was returning the full event no matter
what key was being pressed. The method has now been reimplemented to
return 0 for all focused keypresses. This is probably a little clumsy,
probably only want to kill the shortcut for the relevant keys, which can
be done within the method iff statements...