Restructured SceneLinkServoTabW around ServoDataConnectors
Previously this table generated its items from rows of the
AttributeServoConnector data. This has been upgraded to allow each
AttributeServoConnector to be able to drive multiple servos. As such the
ServoDataConnector class has been created to control how the data connects
to each servo channel.

Each AttributeServoConnector has at least on ServoDataConnector. One as
default, but using a right click on the servo channel column, more can be
added to the AttributeServoConnector.

As such, now all active AttributeServoConnectors are first collected, then
from these all the active DataServoConnectors are collected. The table is
then populated from this list.

Active AttributeServoConnectors are now enforced, but checking whether the
pin is active, and if it is, then passing the node on to be scanned for
AttributeServoConnectors