Tweaked the display of the nodeLinksTableWidget and SkinTableWidget
These tables now clear properly when the happy face is compeletely
cleared. The skinning Table also updates if the current SuperNode is
deleted.

Updates needed to update the nodeLinkdsTableWidget when items are deleted.

To update hte skinning table for superNode deletion, the keypressEvent in
the RigGraphicsView has to check if delete is being pressed, then check
what item type we have, then reach up out through self.mainWindow, to
update the table UI elements. Feels little messy, but it works...