Mirroring of ConstraintItems now works for SuperNodeGroups
reflectSuperNodeGroup method is now extended to mirror any constraintItems
too. Each constraint Item class as a new method called mapAttributes that
allows the attributes of the ConstraintItem to be quickly passed to a new
instance that is created for the mirror. This is then bound to the pin in
the the normal way.

Finally the rotation of the pin itself is measured and the negative is
applied to the mirrored pin.