Mirroring of constrain Items on Wiregroups now works
By applying exactly the same extension to reflect method, we now look
through the new pins in the mirrored wireGroup and copy across and mirror
the constraint Item data.

There is also a tweak that locks down superNodeGroup pins when the group
is dropped from the creation drag. Mirrored superNodeGrop pins are now
also locked down, so they cannot be accidently moved...