Basic addition and check of a Maya Scene Control
New method added to dataProcessor called isSceneControllerActive. This is
run whenever a new superNodeGroup is dragged in, or is a new Wiregroup is
created. If self.sceneControl is found and verified in the Maya scene then
no action is taken. If it is not found, or is None, then the user is asked
to put in a unique name to build a new one.

The sceneAppData is now passed down to the ServoDataConnector level, so
appropriate attributes can be createdon the sceneControl, and connected to
appropirate animation curves nodes that will be created in the Maya scene.

dataprocessor, or sceneControl is going to have to be passed down too.
Investigate this...