WireGroups and SuperNodes now add attributes to the sceneControl
Attributes are currently added with nice locked off boolean labels.
However, they are not checked or removed. to do this the dataProcessor is
passed down through all levels down to the servoDataChannel, but not when
the instance of each class is created, since that failed on a deepcopy.

As such a method in the dataclass has been created called manageAttribute
Connections. This initially collects all the dataBundles. Then ensures
that each one has a the correct dataProcessor passed to it, then loops
through the Attribute connectors adding attributes to the SceneControl
after initially creating a title for each DataBundle.