ServoDataConnectors now manage animCurve Connections
ServoDataConnectors now have a method called manageServoCurveNode. This
checks initially creates a servoCurveNode if none exists. If one does
exist in the scene, then it strips it of connections and reattaches it to
the correct attribute.

Extra methods also added in Maya Data associated with listing, making and
breaking connections