ServoDataConnector animCurves now clean up on Deletion
When nodes are deleted, then now call a clearData method which calls the
same method in the dataBundle and all the way down to servoDataConnector,
cleaning up the data on the way. In the servoDataConnector it checks to
see if the animCurve exists using the deleteServoCurveMethod. If found the
curve Node is deleted from the maya scene.

Now node creation and deletion continually links up the correct attributes
and animCurves for the servoDataConnectors.

Still yet to check is adding and removing individual servo Connectors