Extra ServoDataConnectors can be added and removed with AnimCurves
The servoCurveNodes now fall in line and are deleted, created and renamed
appropriately when extra servoDataConnectors are added and removed from
attribute Connectors.

This is done by addapting the attributeServoConnector
removeServoDataConnector method to delete the associated servoCurveNode.

The sortServoDataConnectors is also expanded to temporarily rename
servoCurveNodes, so there is no clash in Maya names. Then sort the new
indexes and names, then rename the servoCurveNodes to fall back in line.