AttributeConnector setValue updated to read AnimCurve
The setValue method in the AttributeConnector is now updated so the
incoming value is scaled between -10 and 10, and then fed out to the
animCurve to get the final return value, allowing complex relationships
through the curve...

Range increased from -1 to 1, to -10 to 10, because it makes it easier to
insert extra keys in the graph for more control.

Now we need to reimplement setValue for AttributeServoConnectors, so that
the servo data is processed and reported back too.