AttributeServoConnector has new setValue method for servos
Reimplemented method that simply calls the servoDataConnector.setServo for
each servoDataConnector. The sceneControl attribute has already been set,
so it is just a case of getting the output from the associated
servoCurveNodes. Simples..

Next stage is to actually implement the servo moving code. From there
everything has to be wrapped up with proper, read and store methods.