Moving Servo Code implemented through MaestroServo module
Separate module defined with a class called MaestronSerialServo. This can
be swapped out if there is any other servo control mechanism in the
future. This class is fed into the creation of the DataProcessor.

It cannot be included in the DataBundle, since the dataBundle is being
deep copied, which messes up the serial connection, throwing an error.

As such the servoAppData is passed down from DataProcessor, to
DataServoBundle, to AttributeServoConnector to ServoDataConnector using
the setServoAppData method, in exactly the same way as the
setDataProcessor method was implemented. The setServoAppData method is
called as soon as a new bundle is made. Appropriate calls for additional
ServoDataConnectors are also made.

Everything works!

Now implement storage of the DataProcessor and tidy up saving on the
attributes that are now needed.