Random Bug hunt from play testing
From messing around building faces a number of small bugss needed fixing.
The following are fixed:

Color setting of wiregroups referenced old group methods. Updated.

Mirrored controls were not initiated with proper dataBundle information.

Deactivated pins were still giving out move messages on a "Reset
Wiregroup", which was causing an error trying to read a curveNode that had not been built.

On file read the Maestro Servo information was not being passed down the
dataProcessor

On file read the nodeLinks data Table highlighting was failing because the
node attribute had not been set in the AttributeConnector

Test loading of a deactivated pin --> Activating it. Seeing if there is an
error on movement.

Serial needs to be closed down when the program is exited. At the moment
it is not. So the program cannot be relaunched.