Bug Fix for mouse moving out of RigGV on drag and drop
on the RigGV there is now a dragLeaveEvent method that manages if a user
drags an item in the RigGV, but leaves the GV with the mouse before
dropping. Before this caused multiple items to spawn and lots of errors.
Now this method checks what the dragged item is and cleans up accordingly.