Expression Capture started with RigGV Control Item
Basic outline of the ExpressionCaptureNode has been created as an extra
control Item. Variable in size based off its text. Check that resizing
works for different sizes of text. It wont. self.borderRect has to be
passed through the rest of the paint commands and rectangles.

ExpressionSlider class has been developed and is a child of the
ExpressionCaptureNode. Sliding movement works fine and returns a
percentage correctly for its position.

Next develop actual ExpressionCaptureData models.