expresionCapture.py created for Expression Classes
Main Classes are blocked in for Expression Capture.

ExpressionItemState - this records the details of a rigView control Node.
records the node, index, group, groupName and reference Position.

ExpressionFaceState - creates and collects an ExpressionItemState for each
control Node in the scene. Also records an assoicated name for that
expressions and a corresponding ExpressionStateNode from the rigGV.

ExpressionCaptureProcessor - This is the top level of Expressions Capture.
It has a faceSnapShot which is a ExpressionFaceState for the current
position of the face. It then collects all the expressions in an
expressionLibrary list, so comparisions and final node positions can be
calculated from the data.