Expression Capture System is working! No store/Read
The capture system is now full implemented mainly through the
ExpressionCaptureProcessor.processCombinedExpressions method. This method
looks at the ExpressionFaceState in the self.faceSnapShot, and compares
each controller that it finds to the positioning in that faceState. From
there the total difference can be calculated. The correct percentage is
taken of this, and stored as a delta in the faceSnapShot
ExpressionItemStates. Once all the deltas have been added up for all the
expressions, then the final position of the nodes can be calculated, and
the nodes are placed there.

Setting a new faceSnapShot resets all the sliders on all the
ExpressionStateNodes to zero, so no expression is contributing anything.
From there, any expression can be dialled in....