Partial Expression implemented with Selections
The system now records your selection before moving an ExpressionStateNode
slider, and then the slider will only act on the nodes that were selected.
Using the mousePress/ReleaseEvents in the slider the selection of the
rigGV is then restored afterwards.

The expression monitoring includes each ExpressionItemState having a
"selected" attribute. If no nodes are selected then the entire expression
is dialled in across all nodes.

The selection state for each node is collected for the whole view in the
mapSelected method of the ExpressionCaptureProcessor.