ExpressionStateNode now collects colour data from ExpressionFaceState
Attempting to collect all major data in the ExpressionFaceState. This way
only this needs to be saved and loaded, and it can create the appropriate
graphical representation when it does so (creating a new
ExpressionStateNode on read).

Appropirate attributes added to the ExpressonFaceState so that it can
record colour and position information for both the ExpressionStateNode
and its slider.