ExpressionCapture Processor store/read blocked in
The methods read and store have been blocked in without syntax error all
the way down the expression system levels. Very limited testing. RigStore
is updated to try and record the ExpressionSystem in XML.

Initial Problem is that the faceSnapShot is a ExpressionFaceState, but has
no ExpressionStateNode associated with it, so things like getPos() fail
when store() is called. Look at changing store and read behaviour
depending on the existence of an ExpressionStateNode...