Expression System Saves and Loads, but is flawed on Load.
Expression system slider stuff etc is over engineered and built by hand in
a messy way. It turns out that you can use QGraphicsWidget to actually
build widgets directly into the graphics view. Fortunately functionality
is split off, so now I am going to be rewriting the ExpressionStateNode as
a QGraphicsWidget instead, with a QGraphicsLayout with a label and a
slider. This is last stand of the old system...