New ExpressionStateNode implemented in ControlItems
This is the rewrite of the old ExpressionStateNode with its
ExpressionPercentageSlider. The new class subclasses QGraphicsWidget,
which allows you to embed other QWidgets in the QGraphics View. As such
the QGraphicsWidget takes a QLabel and QSlider as inputs and builds a
small box around them with a nice colour background.

Functionality has not had to be adjusted, and at this stage there are
still old methods left in the ExpressionCapture systems mainly
ExpressionFaceStates that are now redundant.

Read and Store is not implemented. This needs to be cleaned up now...