New Expression Capture System now saves and loads
Saving and loading is now implmented (store, read), and works with
multiple expressions. There is a bug where the wiregroup pinTies do not
correctly line up with the nodes. Look into fixing this.

All data now is concerned with the position and setting of the QSlider.
This is much more efficient, than the previous method.